Started on new console/reporter layer.
