--- types/elegant-spinner/index.d.ts
@@ -1,4 +1,0 @@
-declare module 'elegant-spinner' {
-  export const frames: string[];
-  export default function spinner(): () => string;
-}
